Add property JUNIT_HOME
junit.jar and hamcrest-core.jar may be in CLASSPATH in environment. Then
we are fine.
But in nested build environments the location may be determined by upper
build script. By having a property which can explicitely be given as
parameter with ant call any script may define the location of both. And
if not defined the environment CLASSPATH for local production still
works.